+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+### Added
+
+- Added the clear cart button and set the quoteId to 0
+
 ## [1.0.4] - 2022-03-23
 
 ### Fixed